Organize contexts and isolate legacy scripts
