+++ legacy/modules/products.js
