+++ src/components/FavoritesTab.jsx
@@ -1,7 +1,7 @@
 // src/components/FavoritesTab.jsx
 import React, { useEffect, useRef } from "react";
-import { useFavorites } from "../favorites/FavoritesContext";
-import { useAuth } from "../auth/AuthContext";
+import { useFavorites } from "../contexts/FavoritesContext";
+import { useAuth } from "../contexts/AuthContext";
 
 export default function FavoritesTab({
   refreshKey,
