+++ src/components/buttons/LikeButton.jsx
@@ -1,8 +1,8 @@
 // src/components/buttons/LikeButton.jsx
 import React from "react";
 import SvgLike from "../svgs/SvgLike";
-import { useAuth } from "../../auth/AuthContext";
-import { useSidebar } from "../../ui/SidebarContext";
+import { useAuth } from "../../contexts/AuthContext";
+import { useSidebar } from "../../contexts/SidebarContext";
 import { upvoteProduct } from "../../services/voteService";
 
 export default function LikeButton({ itemId, itemTypeName, onSuccess }) {
