+++ src/contexts/FavoritesContext.jsx
@@ -5,7 +5,7 @@ import React, {
   useState,
   useCallback,
 } from "react";
-import { useAuth } from "../auth/AuthContext";
+import { useAuth } from "./AuthContext";
 import { getItemTypeNameFromId } from "../services/voteService";
 
 const VOTES_PRODUCTS = "https://fastapi.edgevideo.ai/tracking/votes/products";
